+++ src/actions/DeleteWithi18nAction.tsx
@@ -39,11 +39,20 @@ export const DeleteWithi18nAction = ({ id, type, draft, published, onComplete }:
       setIsDeleting(true);
       setConfirmDialogOpen(false);
       const client = getSanityClient();
-      deleteOp.execute();
+
+      const baseTransaction = client.transaction();
+      baseTransaction.delete(`drafts.${baseDocumentId}`);
+      baseTransaction.patch(baseDocumentId, {
+        unset: [config.fieldNames.references],
+      });
+      await baseTransaction.commit();
+
       const translatedDocuments = await getTranslationsFor(baseDocumentId);
-      const transaction = client.transaction();
-      translatedDocuments.forEach(doc => transaction.delete(doc._id));
-      await transaction.commit();
+      const translationsTransaction = client.transaction();
+      translatedDocuments.forEach(doc => translationsTransaction.delete(doc._id));
+      await translationsTransaction.commit();
+
+      deleteOp.execute();
       if (onComplete) onComplete();
     } catch (err) {
       toast.push({
@@ -51,6 +60,8 @@ export const DeleteWithi18nAction = ({ id, type, draft, published, onComplete }:
         status: 'error',
         title: err.toString(),
       });
+    } finally {
+      setIsDeleting(true);
     }
   }, [baseDocumentId, deleteOp, onComplete]);