+++ src/structure/components/TranslationsComponentFactory/TranslationsComponentFactory.tsx
@@ -39,10 +39,12 @@ export const TranslationsComponentFactory =
             config.languages,
             draft ?? published
           );
-          const doc = await getSanityClient().fetch("*[_id == $id]", {
-            id: getBaseIdFromId(props.documentId),
+          const baseDocId = getBaseIdFromId(props.documentId)
+          const doc = await getSanityClient().fetch(`coalesce(*[_id == $draftId][0], *[_id == $id][0])`, {
+            id: baseDocId,
+            draftId: `drafts.${baseDocId}`
           });
-          if (doc && doc.length > 0) setBaseDocument(doc[0]);
+          if (doc) setBaseDocument(doc);
           setLanguages(langs);
           setPending(false);
         }